Replace terminal colors with unicode emoji everywhere but in cli
