+++ pogam/cli.py
@@ -7,7 +7,7 @@ import boto3
 import click
 import click_log  # type: ignore
 
-from . import S3_TASKS_FILENAME, SOURCES, color, create_app, scrapers
+from . import S3_TASKS_FILENAME, SOURCES, create_app, scrapers
 from .models import Listing
 
 logger = logging.getLogger("pogam")
@@ -16,6 +16,22 @@ app = create_app("cli")
 
 TRANSACTION_TYPES = ["rent", "buy"]
 PROPERTY_TYPES = ["apartment", "house", "parking", "store"]
+
+
+# terminal colors
+class color:
+    RED = "\033[31m"
+    GREEN = "\033[32m"
+    PURPLE = "\033[95m"
+    CYAN = "\033[96m"
+    DARKCYAN = "\033[36m"
+    BLUE = "\033[94m"
+    LIGHT_GREEN = "\033[92m"
+    YELLOW = "\033[93m"
+    LIGHT_RED = "\033[91m"
+    BOLD = "\033[1m"
+    UNDERLINE = "\033[4m"
+    END = "\033[0m"
 
 
 @click.group()
